Begin revamping iterator setup
